# api/api.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: ElektroGo
  description: "Documentació OpenAPI per a l'API REST desenvolupada a partir de l'aplicació **ElektroGo**. <br> Per tal de poder executar les operacions cal estar conectat a la VPN de la UPC. Totes les operacions requereixen d'una api_key. En cas de no tenir-ne la podeu solicitar a qualsevol membre de l'equip."
servers:
  - url: 'http://10.4.41.58:8080/'
paths:
...
  /drivers/{username}:
    get:
      tags:
      - Drivers
      summary: Obté un conductor.
      operationId: getDriver
      parameters:
      - name: username
        in: path
        description: Username del conductor
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object 
                $ref: '#/components/schemas/Driver'
        432:
          description: 'Error: Driver Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error432'
    post:
      tags:
      - Drivers
      summary: Crea un conductor.
      operationId: createDriver
      parameters:
      - name: username
        in: path
        description: Username del conductor
        required: true
        schema:
          type: string
      responses:
        201:
          description: successful operation
        439:
          description: 'Error: Driver Alredy Exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error439'
    delete:
      tags:
      - Drivers
      summary: Esborra un conductor.
      operationId: deleteDriver
      parameters:
      - name: username
        in: path
        description: Username del conductor
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
        432:
          description: 'Error: Driver Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error432'
...
  /car-pooling:
    get:
      tags:
      - Car Pooling
      summary: Obté tots els trajectes de carpooling.
      operationId: getAllCarPoolings
      parameters:
      - name: order
        in: query
        description: Si es posa a true, el llistat resultant estarà ordenat.
        required: false
        schema:
          type: boolean
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarPooling'
    post:
      tags:
      - Car Pooling
      summary: Crea un trajecte de carpooling.
      operationId: postCarPooling
      requestBody:
        content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CarPooling'
      responses:
        201:
          description: successful operation
        444:
          description: 'Error: Trip Already Exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error444'
  /car-pooling/{id}:
    get:
      tags:
      - Car Pooling
      summary: Obté un trajecte de carpooling.
      operationId: getCarPooling
      parameters:
      - name: id
        in: path
        description: ID del trajecte de carpooling
        required: true
        schema:
          type: number
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/CarPooling'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
    delete:
      tags:
      - Car Pooling
      summary: Esborra un trajecte de carpooling.
      operationId: deleteCarPooling
      parameters:
      - name: id
        in: path
        description: ID del trajecte de carpooling
        required: true
        schema:
          type: number
      responses:
        200:
          description: successful operation
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /car-pooling/{id}/cancel:
    put:
      tags:
      - Car Pooling
      summary: Cancela un trajecte de carpooling.
      operationId: cancelTrip
      parameters:
      - name: id
        in: path
        description: ID del trajecte de carpooling que es vol esborrar.
        required: true
        schema:
          type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/User'
        404:
          description: 'Error: Trip Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /car-pooling/{id}/users:
    get:
      tags:
      - Car Pooling
      summary: Obté tots els usuaris d'un trajecte de carpooling.
      operationId: getCarPoolingUsers
      parameters:
      - name: id
        in: path
        description: ID del trajecte de carpooling.
        required: true
        schema:
          type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/User'
        404:
          description: 'Error: Trip Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /car-pooling/search:
    get:
      tags:
      - Car Pooling
      summary: Cerca trajectes de carpooling segons origen, destí, data i hora.
      operationId: searchCarPooling
      parameters:
      - name: latO
        in: query
        description: Coordenada latitud de l'origen
        required: true
        schema:
          type: number
          format: double
      - name: longO
        in: query
        description: Coordenada longitud de l'origen
        required: true
        schema:
          type: number
          format: double
      - name: latD
        in: query
        description: Coordenada latitud del destí
        required: true
        schema:
          type: number
          format: double
      - name: longD
        in: query
        description: Coordenada longitud del destí
        required: true
        schema:
          type: number
          format: double
      - name: sDate
        in: query
        description: Data a partir de la qual es busca. Si no s'especifica data es busca a partir del dia actual.
        required: false
        schema:
          type: string
          format: date
      - name: sTimeMin
        in: query
        description: Hora d'inici de l'interval de cerca. Si no s'especifica hora d'inici de l'interval ni data, es busca a partir de l'hora actual a tots els dies a partir de l'actual. Si no s'especifica hora d'inici de l'interval però si data, es busca a partir de l'hora actual y la data indicada.
        required: false
        schema:
          type: string
          format: time
      - name: sTimeMax
        in: query
        description: Hora final de l'interval de cerca. Si no s'especifica hora final de l'interval ni data, es busca fins les 23:59 de tots els dies a partir de l'actual. Si no s'especifica hora final de l'interval però si data, es busca fins les 23:59 y la data indicada.
        required: false
        schema:
          type: string
          format: time
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarPooling'
        400:
          description: 'Error: Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
  /car-pooling/byCoord:
    get:
      tags:
      - Car Pooling
      summary: Donades unes coordenades (latitude i longitude) rep la informació de tots els punts de car pooling vigents dins d'un radi indicat per parametre (en km).
      operationId: getCarPoolingByCoord
      parameters:
      - name: latitude
        in: query
        description: Coordenada latitud de l'origen
        required: true
        schema:
          type: number
      - name: longitude
        in: query
        description: Coordenada longitude de l'origen
        required: true
        schema:
          type: number
      - name: Radi
        in: query
        description: Radi de l'àrea a analitzar (en km).
        required: true
        schema:
          type: number
      - name: key
        in: query
        description: api_key per poder fer servir la API.
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarPooling'
        400:
          description: 'Error: Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error401'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
  /car-pooling/from/{username}:
    get:
      tags:
      - Car Pooling
      summary: Obté tots els trajectes de carpooling d'un usuari.
      operationId: getCarPoolingsUser
      parameters:
      - name: username
        in: path
        description: Username de l'usuari
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarPooling'
        437:
          description: 'Error: User Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error437'
  /car-pooling/{id}/from/{username}:
    post:
      tags:
      - Car Pooling
      summary: Afegeix un usuari a un trajecte de carpooling.
      operationId: postCarPoolingsUser
      parameters:
      - name: id
        in: path
        description: ID del trajecte de carpooling
        required: true
        schema:
          type: number
      - name: username
        in: path
        description: Username de l'usuari
        required: true
        schema:
          type: string
      responses:
        201:
          description: successful operation
        437:
          description: 'Error: User Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error437'
        404:
          description: 'Error: Trip Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
        448:
          description: 'Error: UserTrip Already Exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error448'
    delete:
      tags:
      - Car Pooling
      summary: Esborra un usuari d'un trajecte de carpooling.
      operationId: deleteCarPoolingsUser
      parameters:
      - name: id
        in: path
        description: ID del trajecte de carpooling
        required: true
        schema:
          type: number
      - name: username
        in: path
        description: Username de l'usuari
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
        437:
          description: 'Error: User Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error437'
        404:
          description: 'Error: Trip Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
        447:
          description: 'Error: UserTrip Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error447'
...
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
          example: "samito"
        email:
          type: string
          example: "samito33@gmail.com"
        id:
          type: string
          example: "104069839557168305566"
        provider:
          type: string
          example: "GOOGLE"
        name:
          type: string
          example: "Samuel Daniel Foster"
        givenName:
          type: string
          example: "Samuel"
        familyName:
          type: string
          example: "Daniel Foster"
        imageUrl:
          type: string
          example: "https://lh3.googleusercontent.com/a-/AOh14GiIcAvvngHoDsa9CP5KmAgA1CfJglhCJjnjJR-z"
...
    Driver:
      type: object
      properties:
        username:
          type: string
          example: "samito"
        status:
          type: string
          example: "verified"
...
    CarPooling:
      type: object
      properties:
        id:
          type: integer
          example: 1
        startDate:
          type: string
          format: date (yyyy-mm-dd)
          example: "2022-01-20"
        startTime:
          type: string
          format: time 24h
          example: "12:00:00"
        offeredSeats:
          type: integer
          example: 4
        occupiedSeats:
          type: integer
          example: 2
        restrictions:
          type: string
          example: 'No es pot fumar dins del vehícle'
        details:
          type: string
          example: 'Acostumo a conduir bastant ràpid'
        vehicleNumberPlate:
          type: string
          example: "2001MMO"
        origin:
          type: string
          example: "FIB Facultat d'Informàtica de Barcelona, C/Jordi Girona, 1-3, 08034 Barcelona"
        destination:
          type: string
          example: "Estadi Montilivi, Avinguda de Montilivi, 141, 17003 Girona"
        username:
          type: string
          example: "PolB"
        cancelDate: 
          type: string
          format: date (yyyy-mm-dd)
          example: "2022-01-19"
        latitudeOrigin: 
          type: number
          example: 41.3894
        longitudeOrigin:
          type: number
          example: 2.1134
        latitudeDestination: 
          type: number
          example: 41.9611
        longitudeDestination:
          type: number
          example: 2.8279
...
    ErrorResult:
      type: object
      properties:
        status:
          type: integer
          enum: [400, 401, 404]
        error: 
          type: string
          enum: ["Bad Request", "Unauthorized", "Not found"]
        message: 
          type: string
    ErrorResultElektroGo:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        path:
          type: string
  examples:
    error400:
      value:
        status: 400
        error: "Bad Request"
        message: "Missing parameter"
    error401:
      value:
        status: 401
        error: "Unauthorized"
        message: "Unvalid api key"
    error404:
      value:
        status: 404
        error: "Not found"
        message: "There are no trips in the radius specified"
    error432:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 432
        error: "Http Status 432"
        path: "/drivers"
...
    error437:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 437
        error: "Http Status 437"
        path: "/users"
...
    error439:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 439
        error: "Http Status 439"
        path: "/drivers/samito"
...
    error444:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 444
        error: "Http Status 444"
        path: "/users/reports"
...
    error447:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 447
        error: "Http Status 447"
        path: "/users/samito/trips/2"
    error448:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 448
        error: "Http Status 448"
        path: "/users/samito/trips/2"